fts: extract reader/options.rs (BoolMode, Bm25Stats, Bm25SearchOptions)
Phase 1d. Move the query-option types — the default-operator BoolMode,
the Bm25Stats idf-source selector, and the Bm25SearchOptions builder —
into reader/options.rs, re-exported from mod.rs so fts::reader::{BoolMode,
Bm25Stats, Bm25SearchOptions} stays reachable for the crate. Types were
already pub; no widening needed. OpenOptions and the index-metadata types
(NormTable, ColumnMeta, FtsColumnConfig) stay in core — a coupled
open-config/metadata group and NormTable is used by phrase/sink; a
metadata.rs is a possible later step. Pure move.
